LCP Warning when loading logo in appbar.tsx
I was receiving the following error when trying to
load the logo in the appbar.tsx file:

"warn-once.js:16 Image with src "/logo.png" was
detected as the Largest Contentful Paint (LCP).
Please add the "priority" property if this image is above the fold."

Above the fold meaning the image is visible when
the page loads without scrolling.

The added priority property will tell the browser
to load the image as a priority.