Be sure the NavBar is always in the correct position.
When I would restart the server, the navbar would
be no longer be at the top of the page and at 100%
width.

I fixed it to the top of the page and at 100% width.

I also added a 64px margin to the top of the application
via globals.css to account for the fixed navbar.
So no content will be hidden underneath the navbar