fix margin timing for Navbar
Directly call marginTop: 64 on the body in layout.tsx, so that the
so this css is applied immediately.

conditionally render the margin depending on if
there is a user logged in or not.

if there is not a user logged in, then the login
page will be rendered without the navbar, so it should
not have the margin